refactor post_spec
